requirements.txt
# This file is @generated by PDM.
# Please do not edit it manually.

annotated-types==0.6.0
anyio==4.2.0
click==8.1.7
colorama==0.4.6;
fastapi==0.109.0
h11==0.14.0
idna==3.6
pydantic==2.5.3
pydantic-core==2.14.6
sniffio==1.3.0
starlette==0.35.1
typing-extensions==4.9.0
uvicorn==0.27.0
tensorflow==2.15.0
python-multipart==0.0.6
pillow==10.2.0
scikit-learn==1.3
nltk==3.8.1